+++ CHANGELOG.md
@@ -6,3 +6,4 @@
 
 - `yt-dlp` dependency
 - Ability to create new info file
+- Refresh playlist to save video ids